+++ test/integration/tests/Full_Register_Deregister.t.sol
@@ -6,7 +6,6 @@ import "test/integration/User.t.sol";
 import "test/integration/IntegrationChecks.t.sol";
 
 contract Integration_Full_Register_Deregister is IntegrationChecks {
-
     using BitmapUtils for *;
 
     // 1. Register for all quorums by churning old operators
@@ -101,7 +100,7 @@ contract Integration_Full_Register_Deregister is IntegrationChecks {
 
         // 3. Old operators re-register for quorums
         // Note: churnTargets.length == quorums.length, so we do these one at a time
-        for (uint i = 0; i < churnTargets.length; i++) {
+        for (uint256 i = 0; i < churnTargets.length; i++) {
             User churnTarget = churnTargets[i];
             bytes memory quorum = new bytes(1);
             quorum[0] = quorums[i];
@@ -109,7 +108,7 @@ contract Integration_Full_Register_Deregister is IntegrationChecks {
             churnTarget.registerOperator(quorum);
             check_Register_State(churnTarget, quorum);
         }
-        
+
         // 4. Original operator re-registers for all quorums by churning old operators again
         operator.registerOperatorWithChurn(quorums, churnTargets, new bytes(0));
         check_Churned_State({
@@ -140,11 +139,9 @@ contract Integration_Full_Register_Deregister is IntegrationChecks {
 
         // Select some quorums to register using churn, and the rest without churn
         bytes memory churnQuorums = _selectRand(quorums);
-        bytes memory standardQuorums = 
-            quorums
-                .orderedBytesArrayToBitmap()
-                .minus(churnQuorums.orderedBytesArrayToBitmap())
-                .bitmapToBytesArray();
+        bytes memory standardQuorums = quorums.orderedBytesArrayToBitmap().minus(
+            churnQuorums.orderedBytesArrayToBitmap()
+        ).bitmapToBytesArray();
 
         // Select churnable operators in each quorum. If needed, deals/deposits assets
         // for the operator, and deregisters operators from standardQuorums to make room
